HDFSC-22 | Set HADOOP_USER_NAME for simple auth configuration
