+++ demo/common-operations/README.md
@@ -37,7 +37,7 @@ In order to build and run this demo project you'll need:
 * Import the demo project into your workspace via "Anypoint Exchange" or "Import..." from "File" menu.
 * From the Package Explorer view, expand the demo project and open the **mule-app.properties** file. Fill in the values for properties as explained below:
     * **config.nameNodeUri** - hostname of the machine where your HDFS server is deployed.
-    * **config.sysUser** - (Optional) a simple user identity of a client process
+    * **config.username** - (Optional) a simple user identity of a client process
 * Run the Mule Project.
 * From your favorite browser access http://localhost:8090
 * A webpage from which you can play with different HDFS operation is going to be displayed.
