+++ demo/common-operations/src/main/app/common-operations.xml
@@ -12,7 +12,7 @@ http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/cor
 
 
 http://www.mulesoft.org/schema/mule/ee/dw http://www.mulesoft.org/schema/mule/ee/dw/current/dw.xsd">
-    <hdfs:config name="hdfs-conf" nameNodeUri="${config.nameNodeUri}" username="${config.sysUser}" doc:name="HDFS"/>
+    <hdfs:config name="hdfs-conf" nameNodeUri="${config.nameNodeUri}" username="${config.username}" doc:name="HDFS"/>
     <http:listener-config name="HTTP_Listener_Configuration" host="0.0.0.0" port="8090" doc:name="HTTP Listener Configuration"/>
     <flow name="UI_Provider_Flow">
         <http:listener config-ref="HTTP_Listener_Configuration" path="/" doc:name="HTTP"/>
