+++ demo/common-operations/src/main/app/mule-app.properties
@@ -1,2 +1,2 @@
 config.nameNodeUri=
-config.sysUser=
+config.username=
