+++ demo/poll-from-file/src/main/app/mule-app.properties
@@ -1,5 +1,5 @@
 # connection related configs
 config.nameNodeUri=
-config.sysUser=
+config.username=
 # flow related configs
 fileToPollFrom=
